Move set_item_start_time, set_item_duration, and ClampPolicy to stack_item_set.rs
ClampPolicy belongs at the stack level (not track), so it lives alongside the
two methods that use it in a dedicated stack_item_set.rs file.
